clip_exercise/requirements.txt
# --- Core Deep Learning & Vision ---
torch>=2.0.0
torchvision
transformers>=4.49.0
Pillow

# --- Qwen-VL & Spatial Reasoning ---
qwen-vl-utils
accelerate>=0.26.0

# --- Performance Optimization (4-bit AWQ) ---
autoawq
bitsandbytes

# --- Project Utilities ---
# Note: tomllib and shutil are standard libraries in Python 3.11+
# and do not need to be listed here.